reuse MultiModeTaxiModule submodules in (single-mode) DrtModule
