+++ contribs/drt/src/main/java/org/matsim/contrib/drt/run/DrtModule.java
@@ -1,36 +1,13 @@
 package org.matsim.contrib.drt.run;
 
-import java.net.URL;
-
-import org.matsim.api.core.v01.Scenario;
-import org.matsim.contrib.drt.optimizer.depot.DepotFinder;
-import org.matsim.contrib.drt.optimizer.depot.NearestStartLinkAsDepot;
-import org.matsim.contrib.drt.optimizer.insertion.InsertionCostCalculator;
-import org.matsim.contrib.drt.optimizer.rebalancing.NoRebalancingStrategy;
-import org.matsim.contrib.drt.optimizer.rebalancing.RebalancingStrategy;
-import org.matsim.contrib.drt.optimizer.rebalancing.mincostflow.MinCostFlowRebalancingModule;
-import org.matsim.contrib.drt.optimizer.rebalancing.mincostflow.MinCostFlowRebalancingParams;
-import org.matsim.contrib.drt.routing.DefaultAccessEgressStopFinder;
-import org.matsim.contrib.drt.routing.DefaultDrtRouteUpdater;
+import org.matsim.contrib.drt.analysis.DrtModeAnalysisModule;
 import org.matsim.contrib.drt.routing.DrtMainModeIdentifier;
-import org.matsim.contrib.drt.routing.DrtRouteUpdater;
-import org.matsim.contrib.drt.routing.DrtRoutingModule;
-import org.matsim.contrib.drt.routing.StopBasedDrtRoutingModule;
-import org.matsim.contrib.drt.routing.StopBasedDrtRoutingModule.AccessEgressStopFinder;
-import org.matsim.contrib.dvrp.data.Fleet;
-import org.matsim.contrib.dvrp.data.file.FleetProvider;
-import org.matsim.contrib.dvrp.passenger.DefaultPassengerRequestValidator;
-import org.matsim.contrib.dvrp.passenger.PassengerRequestValidator;
 import org.matsim.contrib.dvrp.router.TimeAsTravelDisutility;
-import org.matsim.contrib.dvrp.run.DvrpModes;
-import org.matsim.core.config.ConfigUtils;
+import org.matsim.contrib.dvrp.trafficmonitoring.DvrpTravelDisutilityProvider;
 import org.matsim.core.controler.AbstractModule;
 import org.matsim.core.mobsim.qsim.AbstractQSimModule;
 import org.matsim.core.router.MainModeIdentifier;
 import org.matsim.core.router.costcalculators.TravelDisutilityFactory;
-import org.matsim.core.scenario.ScenarioUtils;
-import org.matsim.pt.transitSchedule.api.TransitSchedule;
-import org.matsim.pt.transitSchedule.api.TransitScheduleReader;
 
 import com.google.inject.Inject;
 
@@ -40,69 +17,23 @@ import com.google.inject.Inject;
  */
 public final class DrtModule extends AbstractModule {
 
-	private final AbstractQSimModule drtQSimModule;
-
 	@Inject
 	private DrtConfigGroup drtCfg;
 
-	public DrtModule() {
-		this(new DrtQSimModule());
-	}
-
-	public DrtModule(AbstractQSimModule drtQSimModule) {
-		this.drtQSimModule = drtQSimModule;
-	}
-
 	@Override
 	public void install() {
-		String mode = drtCfg.getMode();
-		bind(DvrpModes.key(Fleet.class, mode)).toProvider(new FleetProvider(drtCfg.getVehiclesFile()))
-				.asEagerSingleton();
-		bind(Fleet.class).annotatedWith(Drt.class).to(DvrpModes.key(Fleet.class, mode));
-
-		bind(DvrpModes.key(PassengerRequestValidator.class, mode)).to(DefaultPassengerRequestValidator.class);
-		bind(DepotFinder.class).to(NearestStartLinkAsDepot.class);
-		bind(TravelDisutilityFactory.class).annotatedWith(Drt.class).toInstance(TimeAsTravelDisutility::new);
-
-		if (MinCostFlowRebalancingParams.isRebalancingEnabled(drtCfg.getMinCostFlowRebalancing())) {
-			install(new MinCostFlowRebalancingModule());
-		} else {
-			bind(RebalancingStrategy.class).to(NoRebalancingStrategy.class).asEagerSingleton();
-		}
-
-		bind(InsertionCostCalculator.PenaltyCalculator.class).to(drtCfg.isRequestRejection() ?
-				InsertionCostCalculator.RejectSoftConstraintViolations.class :
-				InsertionCostCalculator.DiscourageSoftConstraintViolations.class).asEagerSingleton();
+		install(new DrtModeModule(drtCfg));
+		installQSimModule(new DrtModeQSimModule(drtCfg));
+		install(new DrtModeAnalysisModule(drtCfg));
 
 		bind(MainModeIdentifier.class).to(DrtMainModeIdentifier.class).asEagerSingleton();
 
-		switch (drtCfg.getOperationalScheme()) {
-			case door2door:
-				addRoutingModuleBinding(drtCfg.getMode()).to(DrtRoutingModule.class);//not singleton
-				break;
-
-			case stopbased:
-				bind(TransitSchedule.class).annotatedWith(Drt.class)
-						.toInstance(readTransitSchedule(drtCfg.getTransitStopsFileUrl(getConfig().getContext())));
-
-				bind(DrtRoutingModule.class);//not singleton
-				addRoutingModuleBinding(drtCfg.getMode()).to(StopBasedDrtRoutingModule.class);//not singleton
-				bind(AccessEgressStopFinder.class).to(DefaultAccessEgressStopFinder.class).asEagerSingleton();
-				break;
-
-			default:
-				throw new IllegalStateException();
-		}
-
-		bind(DrtRouteUpdater.class).to(DefaultDrtRouteUpdater.class).asEagerSingleton();
-		addControlerListenerBinding().to(DrtRouteUpdater.class);
-
-		installQSimModule(drtQSimModule);
-	}
-
-	static TransitSchedule readTransitSchedule(URL url) {
-		Scenario scenario = ScenarioUtils.createScenario(ConfigUtils.createConfig());
-		new TransitScheduleReader(scenario).readURL(url);
-		return scenario.getTransitSchedule();
+		bind(TravelDisutilityFactory.class).annotatedWith(Drt.class).toInstance(TimeAsTravelDisutility::new);
+		installQSimModule(new AbstractQSimModule() {
+			@Override
+			protected void configureQSim() {
+				DvrpTravelDisutilityProvider.bindTravelDisutilityForOptimizer(binder(), Drt.class);
+			}
+		});
 	}
 }
